Disallow additionalProperties in most schemas
I'd like to move toward tighter validation so we can have easier
composability of these schemas together. To that end, I'm moving toward
disallowing additionalProperties (updated explicitly in the unit tests)
and consolidating as many of the shared definitions as possible and
using a lot more allOf statements. In this way, it moves the whole thing
toward a more multiple-inheritance / mixin style of DRY definitions.

To do this I've had to flip some of the previous valid examples to
invalid ones. And I've had to define all POSSIBLE properties on things
I want to add additionalProperties: false to.

We can still then limit possible properties to just those we want with
things like the noDocumentIndexing schema. When mixed in to an
objectResult schema via allOf, it disallows use of the page, paragraph,
and sentence properties even though they're allowed in objectResult. I
like this pattern moving forward.