VTN-26951 Add vtn-standard json-schemas for text engine categories (#399)
Adds a number of new json-schema validators for engine categories centered around text analytics and translation.  The full list is:
- concept: Content Classification
- entity: Entity Extraction
- keyword: Keyword Extraction
- language: Language Identification
- media-translated: Document (.txt, .docx, etc.) Translation
- sentiment: Textual Sentiment Analysis
- summary: Textual Summarization
- text: Text Extraction, OCR, Extracted Text Translation, and OCR Translation

Includes relevant refactoring to existing object, transcript, and master json-schemas, examples, and tests.

In general, it changes the approach from a more permissive one where additional properties were allowed and the validation would still pass to a more restrictive one where additional properties are no longer allowed in most parts of the schema.  This makes schemas easier to compose and easier to use definitions as "mixins" in certain ways.
To do this some of the  previous valid examples have been changed to
invalid ones. And I've had to define all POSSIBLE properties on things
I want to add additionalProperties: false to.
We can still then limit possible properties to just those we want with
things like the noDocumentIndexing schema. When mixed in to an
objectResult schema via allOf, it disallows use of the page, paragraph,
and sentence properties even though they're allowed in objectResult.

Adds integration with now for CI

Now includes json-schema json files and examples along with the final build

Changed the schema IDs to be where their proper public location will be on the
docs site (once integrated).

Formats veritone-json-schemas with prettier

Cleans up veritone-json-schemas README