VTN-26012 Remove text indexing from content classification real world example
In the "real world" content classification engines don't usually include
text indexing information for individual insights because they usually
cover the content classification for the document as a whole.

But it's fine if we continue supporting it, so I've moved it to a
secondary valid example.
(The "real world" example is used in our docs)